+++ tests/test_uploads.py
@@ -17,26 +17,36 @@ class TestMultipartFormDataFileUpload:
 
     def test_upload_ok(self, httpbin):
         r = http('--form', '--verbose', 'POST', httpbin.url + '/post',
-                 'test-file@%s' % FILE_PATH_ARG, 'foo=bar')
+                 f'test-file@{FILE_PATH_ARG}', 'foo=bar')
         assert HTTP_OK in r
         assert 'Content-Disposition: form-data; name="foo"' in r
         assert 'Content-Disposition: form-data; name="test-file";' \
-               ' filename="%s"' % os.path.basename(FILE_PATH) in r
+               f' filename="{os.path.basename(FILE_PATH)}"' in r
         assert FILE_CONTENT in r
         assert '"foo": "bar"' in r
         assert 'Content-Type: text/plain' in r
 
     def test_upload_multiple_fields_with_the_same_name(self, httpbin):
         r = http('--form', '--verbose', 'POST', httpbin.url + '/post',
-                 'test-file@%s' % FILE_PATH_ARG,
-                 'test-file@%s' % FILE_PATH_ARG)
+                 f'test-file@{FILE_PATH_ARG}',
+                 f'test-file@{FILE_PATH_ARG}')
         assert HTTP_OK in r
         assert r.count('Content-Disposition: form-data; name="test-file";'
-                       ' filename="%s"' % os.path.basename(FILE_PATH)) == 2
+                       f' filename="{os.path.basename(FILE_PATH)}"') == 2
         # Should be 4, but is 3 because httpbin
         # doesn't seem to support filed field lists
         assert r.count(FILE_CONTENT) in [3, 4]
         assert r.count('Content-Type: text/plain') == 2
+
+    def test_upload_custom_content_type(self, httpbin):
+        r = http('--form', '--verbose', 'POST', httpbin.url + '/post',
+                 f'test-file@{FILE_PATH_ARG};type=image/vnd.microsoft.icon')
+        assert HTTP_OK in r
+        # Content type is stripped from the filename
+        assert 'Content-Disposition: form-data; name="test-file";' \
+               f' filename="{os.path.basename(FILE_PATH)}"' in r
+        assert FILE_CONTENT in r
+        assert 'Content-Type: image/vnd.microsoft.icon' in r
 
 
 class TestRequestBodyFromFilePath: