+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 * Added ``--ciphers`` to allow configuring OpenSSL ciphers (`#870`_).
 * Added support for ``$XDG_CONFIG_HOME`` (`#920`_).
 * Fixed built-in plugins-related circular imports (`#925`_).
+* Fixed custom content types for each multipart uploaded file (`#668`_).
 
 
 `2.1.0`_ (2020-04-18)
@@ -441,3 +442,4 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 .. _#895: https://github.com/jakubroztocil/httpie/issues/895
 .. _#920: https://github.com/jakubroztocil/httpie/issues/920
 .. _#925: https://github.com/jakubroztocil/httpie/issues/925
+.. _#668: https://github.com/jakubroztocil/httpie/issues/668